+++ helpers/PrepareRelease.ps1
@@ -1,3 +1,7 @@
-﻿& $PSScriptRoot\GenerateDocs.ps1
-& $PSScriptRoot\RunPSScriptAnalyzer.ps1
-& $PSScriptRoot\RunPester.ps1
+﻿param(
+    [string] $ModuleName = 'PSTibber'
+)
+
+& $PSScriptRoot\GenerateDocs.ps1 -ModuleName $ModuleName
+& $PSScriptRoot\RunPSScriptAnalyzer.ps1 -ModuleName $ModuleName
+& $PSScriptRoot\RunPester.ps1 -ModuleName $ModuleName