+++ CHANGELOG.md
@@ -1,9 +1,7 @@
 # Changelog
 
-See [`README.md`](README.md) and the documentation available [in `docs`](docs/) for more details.
-
 ## Version 0.1.0
 
 * Initial version.
 * Specific functions for getting user info, home info, power consumtion/production, and price info.
-* Any GraphQL query can be executed using the `Invoke-TibberGraphQLQuery` function.
+  *Note: Any GraphQL query can be executed using the `Invoke-TibberGraphQLQuery` function, see [README.md](README.md#usage).*