+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Version 0.1.0
 
-* Initial version.
-* Specific functions for getting user info, home info, power consumtion/production, and price info.
-  *Note: Any GraphQL query can be executed using the `Invoke-TibberGraphQLQuery` function, see [README.md](README.md#usage).*
+* Initial version, see [README.md](README.md#usage) for how to use this module.
+* Generic function for invoking any GraphQL query, [`Invoke-TibberGraphQLQuery`](docs/functions/Invoke-TibberGraphQLQuery.md).
+* Functions for getting user info, home info, energy price info, and energy consumtion/production data:
+  * [`Get-TibberUser`](docs/functions/Get-TibberUser.md)
+  * [`Get-TibberHome`](docs/functions/Get-TibberHome.md)
+  * [`Get-TibberPriceInfo`](docs/functions/Get-TibberPriceInfo.md)
+  * [`Get-TibberConsumption`](docs/functions/Get-TibberConsumption.md)
+  * [`Get-TibberProduction`](docs/functions/Get-TibberProduction.md)